+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 #### 0.0.1-alpha - May 30 2016
 * Initial setup
+* First result implementation